LayoutTests/inspector-obsolete/timeline/timeline-network-resource-expected.txt

CONSOLE MESSAGE: line 7: Script evaluated
Tests the Timeline API instrumentation of a network resource load

Script resource loaded

ResourceSendRequest Properties:
{
    counters : <object>
    data : {
        requestId : <string>
        requestMethod : "GET"
        url : <string>
    }
    frameId : <string>
    stackTrace : <object>
    startTime : <number>
    type : "ResourceSendRequest"
    usedHeapSize : <number>
}

ResourceReceiveResponse Properties:
{
    children : <object>
    counters : <object>
    data : {
        mimeType : <string>
        requestId : <string>
        statusCode : 0
    }
    endTime : <number>
    frameId : <string>
    startTime : <number>
    type : "ResourceReceiveResponse"
    usedHeapSize : <number>
}

ResourceFinish Properties:
{
    counters : <object>
    data : {
        didFail : false
        requestId : <string>
    }
    frameId : <string>
    startTime : <number>
    type : "ResourceFinish"
    usedHeapSize : <number>
}
